shared_config_with_keyring: Create .aws directory if it doesn't exist

+++ lib/aws_assume_role/store/shared_config_with_keyring.rb
@@ -225,9 +225,13 @@ class AwsAssumeRole::Store::SharedConfigWithKeyring < AwsAssumeRole::Vendored::A
 
     # Please run in a mutex
     def save_configuration
-        bytes_required = File.size(determine_config_path)
-        random_bytes = SecureRandom.random_bytes(bytes_required)
-        File.write(determine_config_path, random_bytes)
+        if File.exist? determine_config_path
+            bytes_required = File.size(determine_config_path)
+            random_bytes = SecureRandom.random_bytes(bytes_required)
+            File.write(determine_config_path, random_bytes)
+        else
+            FileUtils.mkdir_p(Pathname.new(determine_config_path).dirname)
+        end
         configuration.save
     end
 end
